Move `step` to a separate module
