+++ packages/next-playwright/src/index.ts
@@ -1,3 +1,5 @@
+import { step } from './step'
+
 /**
  * Minimal interfaces for Playwright's Page and BrowserContext. We use
  * structural types rather than importing from a specific Playwright package
@@ -21,32 +23,6 @@ interface PlaywrightPage {
   url(): string
   context(): PlaywrightBrowserContext
 }
-
-type Step = <T>(title: string, body: () => Promise<T>) => Promise<T>
-
-let step: Step = (_title, body) => body()
-try {
-  const pw = require('@playwright/test') as typeof import('@playwright/test')
-  if (typeof pw.test?.step === 'function') {
-    const playwrightStep = pw.test.step
-    step = async (title, body) => {
-      try {
-        return await playwrightStep(title, body)
-      } catch (e) {
-        // If test.step fails because we're not inside the Playwright test
-        // runner (e.g., running under Jest), fall back to executing the body
-        // directly without step labels.
-        if (
-          e instanceof Error &&
-          e.message.includes('can only be called from a test')
-        ) {
-          return body()
-        }
-        throw e
-      }
-    }
-  }
-} catch {}
 
 const INSTANT_COOKIE = 'next-instant-navigation-testing'
